+++ Gunner/Assets/__Scripts/Health/HealthEvent.cs
@@ -8,7 +8,7 @@ public class HealthEvent : MonoBehaviour
 {
     public event Action<HealthEvent, HealthEventArgs> OnHealthChanged;
 
-    public void CallHealthChangedEvent(float healthPercent, int healthAmount, int damageAmount)
+    public void CallHealthChangedEvent(float healthPercent, float healthAmount, float damageAmount)
     {
         OnHealthChanged?.Invoke(this, new HealthEventArgs() { healthPercent = healthPercent, healthAmount = healthAmount,
             damageAmount = damageAmount });
@@ -18,6 +18,6 @@ public class HealthEvent : MonoBehaviour
 public class HealthEventArgs : EventArgs
 {
     public float healthPercent;
-    public int healthAmount;
-    public int damageAmount;
+    public float healthAmount;
+    public float damageAmount;
 }
